Changed Font::getHeight to return a float like it should. Added minimum message height to GuiMsgBox. TextComponent now vertically centers text. Fixed a bug that would cause ScraperSearchComponent to return results continuously until another search was started.

+++ src/components/TextEditComponent.cpp
@@ -206,7 +206,7 @@ void TextEditComponent::render(const Eigen::Affine3f& parentTrans)
 			cursorPos[1] = 0;
 		}
 
-		Renderer::drawRect((int)cursorPos.x(), (int)cursorPos.y(), 3, f->getHeight(), 0x000000FF);
+		Renderer::drawRect((int)cursorPos.x(), (int)cursorPos.y(), 3, (int)f->getHeight(), 0x000000FF);
 	}
 
 	Renderer::popClipRect();
